New !ytmix and !ytmusicmix commands to play YouTube mixes (#47)
New commands to play YouTube auto-generated mix playlists for videos and
songs. This can be done easily thanks to YouTube always using the the
same url format for the mixes:

`https://www.youtube.com/watch?v={video_id}&list=RD{video_id}`

If a search query is used it'll be searched and the mix of the first
result will be played.

This will obviously only work with single YouTube videos or songs, not
with playlists.