YouTube Music mix (this time for real)
Previously the YouTube Music mix command searched the song on YouTube Music and started the non-music YouTube mix for that song, which included music videos. Now we use another library to properly get the YouTube Music radio playlist for the song.

Also remove "Obtaining information..." message to avoid spamming it with these mixes, it's not needed anymore with slash commands.