Add new HuggingChat implementation for /ask command
OpenAI doesn't give free credits to new accounts anymore, so that's a big no-no from us. Use hugchat library to generate responses to /ask command, which uses HuggingChat API from HuggingFace. An user and password from HuggingFace are required to be added to keys.json.

Also, instead of the old OpenAiClient implementation a new interface is used in the rest of the code, so changing the implementation in the future is easier.